can skip fault localization

+++ src/main/java/mfix/core/locator/SBFLocator.java
@@ -45,11 +45,13 @@ public class SBFLocator extends AbstractFaultLocator {
     @Override
     protected void locateFault(double threshold) {
         try {
-            LevelLogger.info("Perform SBFL ....");
-            Utils.deleteFiles(getBuggyLineSuspFile());
-            ExecuteCommand.execute(CmdFactory.createSbflCmd((D4jSubject) _subject, Constant.SBFL_TIMEOUT),
-                    _subject.getJDKHome(), Constant.D4J_HOME);
-            LevelLogger.info("Finish SBFL ...");
+            if (!Constant.SKIP_FAULTLOCALIZATION) {
+                LevelLogger.info("Perform SBFL ....");
+                Utils.deleteFiles(getBuggyLineSuspFile());
+                ExecuteCommand.execute(CmdFactory.createSbflCmd((D4jSubject) _subject, Constant.SBFL_TIMEOUT),
+                        _subject.getJDKHome(), Constant.D4J_HOME);
+                LevelLogger.info("Finish SBFL ...");
+            }
         } catch (Exception e) {
             e.printStackTrace();
         }
@@ -77,7 +79,10 @@ public class SBFLocator extends AbstractFaultLocator {
 
     @Override
     public List<Location> getLocations(int topK) {
-        List<Location> lines = getSortedSuspStmt(getBuggyLineSuspFile(), topK);
+        List<Location> lines = null;
+        if (!Constant.SKIP_FAULTLOCALIZATION) {
+            lines = getSortedSuspStmt(getBuggyLineSuspFile(), topK);
+        }
         if (lines == null || lines.isEmpty()) {
             lines = ochiaiResult(getOchiaiFile(), topK);
         }
